Revise Step 4-2 markdown gif file - gif 설명에 백그라운드 타이머, 이벤트 루프 추가 - 큐에 setTimeout이 아니라 익명 함수가 들어간다는 것을 명시적인 그림으로 교체

+++ cs23/step2/alu.h
@@ -22,12 +22,12 @@ public:
 
     Word sub_impl(const Reg& r1, const Reg& r2)
     {
-        return r1.get_value() - r2.get_value();
+        return (int16_t )r1.get_value() - (int16_t )r2.get_value();
     }
 
     Word sub_impl(const Reg& r1, const Word value)
     {
-        return r1.get_value() - value;
+        return (int16_t )r1.get_value() - (int16_t )value;
     }
     
     Word and_impl(const Reg& r1, const Reg& r2)
